feat: Add `try_finally` helper.
This implements a pattern for simulating a classic "try/finally" block
over some async Result-returning code. It's particularly helpful for
testcases that need to execute some async cleanup code at the end of
a test run, regardless of whether the test succeeded, failed, or panicked.
